dummy-test-api/requirements.txt
-i https://pypi.org/simple
click==7.1.2
flask==1.1.2
itsdangerous==1.1.0
jinja2==2.11.2
markupsafe==1.1.1
python-dotenv==0.13.0
werkzeug==1.0.1
